requirements.txt
Flask==1.1.1
numpy==1.17.2
spacy==2.2.1
Werkzeug==0.16.0
scikit_learn==0.21.3
gunicorn==19.9.0
boto3==1.10.0
requests==2.23.0
pdfminer.six==20181108
python-dotenv==0.12.0